Fix warning: implicit conversion changes signedness
visp/modules/core/src/math/robust/vpRobust.cpp: In member function ‘int vpRobust::partition(vpColVector&, int, int)’:
visp/modules/core/src/math/robust/vpRobust.cpp:265:16: warning: conversion to ‘unsigned int’ from ‘std::size_t’ {aka ‘long unsigned int’} may change the sign of the result [-Wsign-conversion]
  265 |   double v = a[static_cast<std::size_t>(r)];
      |                ^~~~~~~~~~~~~~~~~~~~~~~~~~~
visp/modules/core/src/math/robust/vpRobust.cpp:268:14: warning: conversion to ‘unsigned int’ from ‘std::size_t’ {aka ‘long unsigned int’} may change the sign of the result [-Wsign-conversion]
  268 |     while (a[static_cast<std::size_t>(++i)] < v) { }
      |              ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~
visp/modules/core/src/math/robust/vpRobust.cpp:270:18: warning: conversion to ‘unsigned int’ from ‘std::size_t’ {aka ‘long unsigned int’} may change the sign of the result [-Wsign-conversion]
  270 |     while (v < a[static_cast<std::size_t>(--j)]) {
      |                  ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~
visp/modules/core/src/math/robust/vpRobust.cpp:278:17: warning: conversion to ‘unsigned int’ from ‘std::size_t’ {aka ‘long unsigned int’} may change the sign of the result [-Wsign-conversion]
  278 |     std::swap(a[static_cast<std::size_t>(i)], a[static_cast<std::size_t>(j)]);
      |                 ^~~~~~~~~~~~~~~~~~~~~~~~~~~
visp/modules/core/src/math/robust/vpRobust.cpp:278:49: warning: conversion to ‘unsigned int’ from ‘std::size_t’ {aka ‘long unsigned int’} may change the sign of the result [-Wsign-conversion]
  278 |     std::swap(a[static_cast<std::size_t>(i)], a[static_cast<std::size_t>(j)]);
      |                                                 ^~~~~~~~~~~~~~~~~~~~~~~~~~~
visp/modules/core/src/math/robust/vpRobust.cpp:280:15: warning: conversion to ‘unsigned int’ from ‘std::size_t’ {aka ‘long unsigned int’} may change the sign of the result [-Wsign-conversion]
  280 |   std::swap(a[static_cast<std::size_t>(i)], a[static_cast<std::size_t>(r)]);
      |               ^~~~~~~~~~~~~~~~~~~~~~~~~~~
visp/modules/core/src/math/robust/vpRobust.cpp:280:47: warning: conversion to ‘unsigned int’ from ‘std::size_t’ {aka ‘long unsigned int’} may change the sign of the result [-Wsign-conversion]
  280 |   std::swap(a[static_cast<std::size_t>(i)], a[static_cast<std::size_t>(r)]);
      |                                               ^~~~~~~~~~~~~~~~~~~~~~~~~~~
visp/modules/core/src/math/robust/vpRobust.cpp: In member function ‘double vpRobust::select(vpColVector&, int, int, int)’:
visp/modules/core/src/math/robust/vpRobust.cpp:302:12: warning: conversion to ‘unsigned int’ from ‘std::size_t’ {aka ‘long unsigned int’} may change the sign of the result [-Wsign-conversion]
  302 |   return a[static_cast<std::size_t>(k)];
      |            ^~~~~~~~~~~~~~~~~~~~~~~~~~~

+++ modules/core/src/math/robust/vpRobust.cpp
@@ -262,12 +262,12 @@ int vpRobust::partition(vpColVector &a, int l, int r)
 {
   int i = l - 1;
   int j = r;
-  double v = a[static_cast<std::size_t>(r)];
+  double v = a[static_cast<unsigned int>(r)];
 
   for (;;) {
-    while (a[static_cast<std::size_t>(++i)] < v) { }
+    while (a[static_cast<unsigned int>(++i)] < v) { }
 
-    while (v < a[static_cast<std::size_t>(--j)]) {
+    while (v < a[static_cast<unsigned int>(--j)]) {
       if (j == l) {
         break;
       }
@@ -275,9 +275,9 @@ int vpRobust::partition(vpColVector &a, int l, int r)
     if (i >= j) {
       break;
     }
-    std::swap(a[static_cast<std::size_t>(i)], a[static_cast<std::size_t>(j)]);
+    std::swap(a[static_cast<unsigned int>(i)], a[static_cast<unsigned int>(j)]);
   }
-  std::swap(a[static_cast<std::size_t>(i)], a[static_cast<std::size_t>(r)]);
+  std::swap(a[static_cast<unsigned int>(i)], a[static_cast<unsigned int>(r)]);
   return i;
 }
 
@@ -299,7 +299,7 @@ double vpRobust::select(vpColVector &a, int l, int r, int k)
       l = i + 1;
     }
   }
-  return a[static_cast<std::size_t>(k)];
+  return a[static_cast<unsigned int>(k)];
 }
 
 /**********************
